Guard pollPipelines w/ new mutex & mv to poller.go
